Refactor: extract fallback/behavior-loss emitters into Diagnostics module (#242 slice 4)
Behavior-preserving extraction of the per-element fallback emission concern
out of HtmlTransformer into Diagnostics/FallbackEmitter:

- capture*Fallback family (html_inline_svg_fallback / html_unsafe_inline_svg,
  html_canvas_runtime_fallback, html_script_fallback,
  html_template_runtime_fallback / html_template_metadata, script_static_metadata)
- runtime-island construction (recordRuntimeIsland -> preserved_runtime_island)
  and required-scripts metadata
- the fallback-only helper subtree (safe*Attributes, scriptRole,
  staticScriptMetadataRole + scriptBody* predicates, templateRole,
  templateRequiresRuntimePreservation)

HtmlTransformer keeps thin delegating wrappers so all call sites are
unchanged; accumulators (runtimeIslands/scriptMetadata) stay transformer-owned
and are passed by reference. sourceContext enrichment is injected as a resolver
closure (it spans the DOM-classification subsystem, not the fallback concern);
small shared leaves (eventMetadata, isSafeSvgContent, dedupeArrayRows,
runtimeIslandSelector) are duplicated per the slice-3 pattern.

Deferred (too entangled, documented): interactive_control_behavior_lost
emission and the inline html_form_fallback / html_iframe_embed_fallback build
literals, which are woven into the interactive-control/form/embed conversion
paths (their recordRuntimeIsland calls already route through the emitter).

Pure move: canonical contracts green, parity 128->128 identical, full
composer test green. HtmlTransformer 5706 -> 5351 lines.
